Fix logger warning

+++ src/Logger.cpp
@@ -28,7 +28,7 @@ void Logger::init(LogLevel logLevel, const std::string& name) {
 	{
 		// Add a file sink
 		char logPath[256];
-		std::snprintf(logPath, sizeof(logPath), "Logs/%s-%lld.log", name.c_str(), std::time(nullptr));
+		std::snprintf(logPath, sizeof(logPath), "Logs/%s-%lld.log", name.c_str(), (long long)std::time(nullptr));
 
 		auto sink = std::make_shared<spdlog::sinks::basic_file_sink_mt>(logPath, true);
 		sink->set_pattern("[%T] [%l] [%@]: %v");
